Dealing with Outliers and Missing data
Outliers and Missing Data: Updating to get more out those data.

For the email data I filled those with mean of poi or non poi depending on their classes and I also filled with 0 for the financial data. I  also analysed more the data and get more sense of the outliers.